two_cylinder_engine_scene.cpp is now headerless
