--- src/Samples/include/babylon/samples/loaders/gltf/morecomplexmodels/2cylinder_engine_scene.h
@@ -1,30 +1,0 @@
-#ifndef BABYLON_SAMPLES_LOADERS_GLTF_MORE_COMPLEX_MODELS_2CYLINDER_ENGINE_SCENE_H
-#define BABYLON_SAMPLES_LOADERS_GLTF_MORE_COMPLEX_MODELS_2CYLINDER_ENGINE_SCENE_H
-
-#include <babylon/interfaces/irenderable_scene.h>
-
-namespace BABYLON {
-namespace Samples {
-
-/**
- * @brief Two Cylinder Engine Scene (glTF). Small CAD data set, including
- * hierarchy.
- * @see https://doc.babylonjs.com/how_to/load_from_any_file_type
- * @see
- * https://github.com/KhronosGroup/glTF-Sample-Models/tree/master/2.0/2CylinderEngine
- */
-struct TwoCylinderEngineScene : public IRenderableScene {
-
-  TwoCylinderEngineScene(ICanvas* iCanvas);
-  ~TwoCylinderEngineScene() override;
-
-  const char* getName() override;
-  void initializeScene(ICanvas* canvas, Scene* scene) override;
-
-}; // end of struct 2CylinderEngineScene
-
-} // end of namespace Samples
-} // end of namespace BABYLON
-
-#endif // end of
-       // BABYLON_SAMPLES_LOADERS_GLTF_MORE_COMPLEX_MODELS_2CYLINDER_ENGINE_SCENE_H
